+++ app/src/main/kotlin/com/urunkarpm/pingpin/ui/settings/SettingsScreen.kt
@@ -80,9 +80,9 @@ fun SettingsScreen(
 
     val currentAppVersion = remember {
         try {
-            context.packageManager.getPackageInfo(context.packageName, 0).versionName ?: "2.1.0"
+            context.packageManager.getPackageInfo(context.packageName, 0).versionName ?: "2.2.0"
         } catch (e: Exception) {
-            "2.1.0"
+            "2.2.0"
         }
     }